wip: consolidate source module
