+++ src/routing/registration/source/normalize-route-sources.ts
@@ -1,5 +1,5 @@
 import type { RouteGroupDefinition } from '@/routing/group/route-group';
-import { mergeRouteOptions } from '@/routing/route/resolve-route-options';
+import { mergeRouteOptions } from '@/routing/registration/source/resolve-route-options';
 import type { RouteDefinition } from '@/routing/route/route-definition.type';
 import type { Request } from 'koa';
 import type { RouteSource } from './route-source';
